refactor(test): update import path from #browser-test to #test/browser
