+++ packages/react/src/components/list/list.test.tsx
@@ -1,4 +1,4 @@
-import { page, render } from "#browser-test"
+import { page, render } from "#test/browser"
 import { BoxIcon } from "../icon"
 import { List } from "./"
